Move tests associated with getting a judgment and its versions into their own file
